Update QEMU version to 2.12 from Debian Buster
Previously, newer qemu package from Ubuntu was downloaded and
installed through dpkg. This is however no longer possible - the
package for QEMU 2.12 depends on a new feature in binfmt-support:
https://bugs.debian.org/868030

The feature needs a corresponding kernel counterpart, so it is not
possible to simply update the binfmt-support package. Travis
workers run on Ubuntu Trusty kernel 4.4, which is not recent enough.

The problem is localized only to the Debian postinst script,
the qemu-*-static binaries still work fine. For this reason,
this commit installs the vendor-provided qemu-user-static
and its packaging files. Then, it manually overwrites the qemu-*-static
executables with the newer ones from qemu 2.12 package.

The QEMU package itself is now taken from Debian Buster repository.